+++ src/main.js
@@ -36,6 +36,7 @@ import collapsePullRequestDescription from './collapse-pull-request-description'
 import setStickyHeader from './sticky-header'
 import setLineLengthLimit from './limit-line-length'
 import collapsePullRequestSideMenus from './collapse-pull-request-side-menus'
+import insertDashboardOverviewFilters from './dashboard-pull-requests'
 
 import observeForWordDiffs from './observe-for-word-diffs'
 
@@ -47,6 +48,7 @@ import {
     isBranch,
     isComparePage,
     isDashBoardOverview,
+    isDashBoardPullRequests,
 } from './page-detect'
 
 import addStyleToPage from './add-style'
@@ -88,6 +90,10 @@ function init(config) {
             comparePagePullRequest()
         }
         codeReviewFeatures(config)
+    } else if (isDashBoardPullRequests()) {
+        if (config.insertDashboardOverviewFilters) {
+            insertDashboardOverviewFilters()
+        }
     }
 
     if (config.addSidebarCounters) {