+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 3.X.X (upcoming)
+
+### Features
+
+-   Add Dashboard Your Work filters into Dashboard Pull Requests page view,
+    [pull request #339](https://github.com/refined-bitbucket/refined-bitbucket/pull/339)
+
 # 3.20.0 (2019-03-01)
 
 ### Features